pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
